SDCICD-257. Add in health check for certman certs.
Certman cert issuance is now part of osde2e's extended health checks
after cluster provisioning. Additionally, metadata has been added to
track this.
