+++ pkg/common/osd/healthchecks/clusterversionoperator_test.go
@@ -1,21 +1,20 @@
-package helper
+package healthchecks
 
 import (
 	"testing"
 
 	configv1 "github.com/openshift/api/config/v1"
 	fakeConfig "github.com/openshift/client-go/config/clientset/versioned/fake"
-	"github.com/openshift/osde2e/pkg/common/config"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
 
-func clusterOperator(name string) *configv1.ClusterOperator {
-	return &configv1.ClusterOperator{
+func clusterVersion() *configv1.ClusterVersion {
+	return &configv1.ClusterVersion{
 		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
+			Name: "version",
 		},
-		Status: configv1.ClusterOperatorStatus{
+		Status: configv1.ClusterVersionStatus{
 			Conditions: []configv1.ClusterOperatorStatusCondition{
 				{
 					Type:    configv1.OperatorAvailable,
@@ -35,13 +34,25 @@ func clusterOperator(name string) *configv1.ClusterOperator {
 					Reason:  "Available",
 					Message: "Available",
 				},
+				{
+					Type:    configv1.OperatorUpgradeable,
+					Status:  configv1.ConditionTrue,
+					Reason:  "Available",
+					Message: "Available",
+				},
+				{
+					Type:    configv1.RetrievedUpdates,
+					Status:  configv1.ConditionTrue,
+					Reason:  "Available",
+					Message: "Available",
+				},
 			},
 		},
 	}
 }
 
-func unavailableClusterOperator(name string) *configv1.ClusterOperator {
-	op := clusterOperator(name)
+func unavailableClusterVersion() *configv1.ClusterVersion {
+	op := clusterVersion()
 	op.Status.Conditions[0].Status = configv1.ConditionFalse
 	op.Status.Conditions[2].Status = configv1.ConditionTrue
 	op.Status.Conditions[0].Message = "Degraded"
@@ -50,8 +61,8 @@ func unavailableClusterOperator(name string) *configv1.ClusterOperator {
 	return op
 }
 
-func progressingClusterOperator(name string) *configv1.ClusterOperator {
-	op := clusterOperator(name)
+func progressingClusterVersion() *configv1.ClusterVersion {
+	op := clusterVersion()
 	op.Status.Conditions[0].Status = configv1.ConditionTrue
 	op.Status.Conditions[1].Status = configv1.ConditionTrue
 	op.Status.Conditions[0].Message = "Available"
@@ -59,39 +70,21 @@ func progressingClusterOperator(name string) *configv1.ClusterOperator {
 	return op
 }
 
-func TestCheckOperatorReadiness(t *testing.T) {
+func TestCheckCVOReadiness(t *testing.T) {
 	var tests = []struct {
 		description string
 		expected    bool
 		objs        []runtime.Object
-		skip        string
 	}{
-		{"no operators", false, nil, ""},
-		{"single operator success", true, []runtime.Object{clusterOperator("a")}, ""},
-		{"single operator failure", false, []runtime.Object{unavailableClusterOperator("a")}, ""},
-		{"single operator progressing", false, []runtime.Object{progressingClusterOperator("a")}, ""},
-		{"multi operator success", true, []runtime.Object{clusterOperator("a"), clusterOperator("b")}, ""},
-		{"multi operator one progressing", false, []runtime.Object{clusterOperator("a"), progressingClusterOperator("b")}, ""},
-		{"multi operator one failure", false, []runtime.Object{clusterOperator("a"), unavailableClusterOperator("b")}, ""},
-		{"multi operator, skip success", true, []runtime.Object{
-			clusterOperator("a"),
-			unavailableClusterOperator("b"),
-			unavailableClusterOperator("c"),
-			unavailableClusterOperator("d"),
-			clusterOperator("e"),
-		}, "b,c,d"},
-		{"multi operator, skip failure", false, []runtime.Object{
-			clusterOperator("a"),
-			unavailableClusterOperator("b"),
-			unavailableClusterOperator("c"),
-			unavailableClusterOperator("d"),
-		}, "b,c"},
+		{"no version", false, nil},
+		{"single version success", true, []runtime.Object{clusterVersion()}},
+		{"single version failure", false, []runtime.Object{unavailableClusterVersion()}},
+		{"single version progressing", false, []runtime.Object{progressingClusterVersion()}},
 	}
 
 	for _, test := range tests {
 		cfgClient := fakeConfig.NewSimpleClientset(test.objs...)
-		config.Instance.Tests.OperatorSkip = test.skip
-		state, err := CheckOperatorReadiness(cfgClient.ConfigV1())
+		state, err := CheckCVOReadiness(cfgClient.ConfigV1())
 
 		if err != nil {
 			t.Errorf("Unexpected error: %s", err)
