+++ pkg/common/osd/healthchecks/nodes_test.go
@@ -1,4 +1,4 @@
-package helper
+package healthchecks
 
 import (
 	"testing"
