+++ pkg/common/osd/healthchecks/pods_test.go
@@ -1,4 +1,4 @@
-package helper
+package healthchecks
 
 import (
 	"testing"
