+++ pkg/e2e/workloads/guestbook/guestbook.go
@@ -9,6 +9,7 @@ import (
 	. "github.com/onsi/gomega"
 
 	"github.com/openshift/osde2e/pkg/common/helper"
+	"github.com/openshift/osde2e/pkg/common/osd/healthchecks"
 	"k8s.io/apimachinery/pkg/util/wait"
 )
 
@@ -45,7 +46,7 @@ var _ = ginkgo.Describe("[Suite: e2e] Workload ("+workloadName+")", func() {
 			// Wait for all pods to come up healthy
 			err = wait.PollImmediate(15*time.Second, 5*time.Minute, func() (bool, error) {
 				// This is pretty basic. Are all the pods up? Cool.
-				if check, err := helper.CheckPodHealth(h.Kube().CoreV1()); !check || err != nil {
+				if check, err := healthchecks.CheckPodHealth(h.Kube().CoreV1()); !check || err != nil {
 					return false, nil
 				}
 				return true, nil
